I did a thing !?!?!
Added single wire(for now)
The wire can connect to two outputs and checks for when its connected to one output or the other 
Wire is capable to detecting when its attacked to both at once and when it disconnects as well.
Wire is also able to move along the 2d space and control its rotation with the left shift key and the w key. 
I can color code it and smooth it out as we go along but the very basics of the wire is complete. The wire does extend on both sides.
Also I moved the code from movePipe into wireScript